backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.3.0
fastapi==0.116.2
h11==0.16.0
idna==3.10
passlib[bcrypt]==1.7.4
pydantic==2.11.9
pydantic_core==2.33.2
pydantic-settings==2.6.1
PyMySQL==1.1.2
python-dotenv==1.1.1
python-jose[cryptography]==3.3.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
